+++ src/api/documentsApi.ts
@@ -1,49 +1,7 @@
 import Api from "./api";
 import notificationLogic from '../components/Notifications/Notification';
 import { successfulCreateAction, successfulEditAction, successfulDeleteAction } from "../components/Notifications/Messages"
-
-export type Document = {
-    id: number;
-    name: string;
-    type: string;
-    organization: string;
-    description: string;
-    date: string;
-    fileName: string | null;
-}
-
-export type MethodicDocumentType = {
-    text: string;
-    value: string;
-}
-export type DocumentOnCreateData = {
-    organizations: Organization[];
-    methodicDocumentTypesItems: MethodicDocumentType[];
-}
-
-export type FileWrapper = {
-    FileAsBase64: string | null;
-    FileName: string | null;
-}
-
-export type Organization = {
-    id: number;
-    organizationName: string;
-}
-
-export type DocumentWrapper = {
-    MethodicDocument: DocumentPost;
-    fileAsBase64: string | null;
-}
-export type DocumentPost = {
-    id: number;
-    name: string;
-    type: number;
-    organization: Organization;
-    description: string;
-    date: string;
-    fileName: string | null;
-}
+import { DocumentOnCreateData, DocumentPost, MethodicDocumentType } from "../models/Documents/DocumentModels";
 
 const dataURLtoFile = (dataurl: string, filename: string) => {
     const arr = dataurl.split(',');
