+++ src/pages/Documents/DocumentsTable.tsx
@@ -2,7 +2,8 @@ import React, { useEffect, useState } from 'react';
 import { Table, Input, Button, Layout, Pagination, Card } from 'antd';
 import columns from './columns';
 import DropDown from './DropDownDocuments';
-import documentsApi, { Document, TypeGetParser } from '../../api/documentsApi';
+import documentsApi, { TypeGetParser } from '../../api/documentsApi';
+import { Document } from "../../models/Documents/DocumentModels";
 import notificationLogic from '../../components/Notifications/Notification';
 import ClickAwayListener from 'react-click-away-listener';
 import moment from "moment";
